[TRY TO FIX] LONG POST

+++ srcs/Config/utils.hpp
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 15:27:37 by tsannie           #+#    #+#             */
-/*   Updated: 2022/01/19 11:05:18 by user42           ###   ########.fr       */
+/*   Updated: 2022/02/08 13:37:06 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ void			checkNbArg( size_t const & toCheck, size_t const & size,
 					std::string const & name );
 void			checkNbArgMin( size_t const & toCheck, size_t const & size,
 					std::string const & name );
-unsigned int	stoui_size( unsigned int const & min, unsigned int const & max,
+unsigned int	stoui_size(	size_t const & min, size_t const & max,
 					std::string const & nb, std::string const & name );
 
 std::string		int_to_string(int x);
